\name{WWR-package}
\alias{WWR-package}
\alias{wwr}
\docType{package}
\title{
\packageTitle{WWR}
}
\description{
\packageDescription{WWR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{WWR}
\packageIndices{WWR}
}
\author{
\packageAuthor{WWR}

Maintainer: \packageMaintainer{WWR}
}
\references{
Pocock S.J., Ariti C.A., Collier T. J. and Wang D. 2012. The win ratio: a new approach to the analysis of composite endpoints in clinical trials based on clinical priorities. European Heart Journal, 33, 176-182.

Luo X., Tian H., Mohanty S. and Tsai W.-Y. 2015. An alternative approach to confidence interval estimation for the win ratio statistic. Biometrics, 71, 139-145.

Bebu I. and Lachin J.M. 2016. Large sample inference for a win ratio analysis of a composite outcome based on prioritized components. Biostatistics, 17, 178-187.

Luo X., Qiu J., Bai S. and Tian H. 2017. Weighted win loss approach for analyzing prioritized outcomes. Statistics in Medicine, <doi: 10.1002/sim.7284>.
}
\keyword{win ratio}
\keyword{weighted win ratio}
\keyword{composite endpoint}
\keyword{pairwise comparison}

